Fixed default songs and poems adding
Due to many recent fixes and changes, the default songs/poems adding
stopped working properly. Reason was that the UNIQUE constraint was not
supposed to be in tables Song and Poem.

So, the first UNIQUE constraint is in table Role, then it has to be in
table User. This fix removed the constraint from Song and Poem. If it
doesn't work, I must remove all the UNIQUE constraints.