Fixed UNIQUE constraints
There was a big issue with tables. One could only add one account after
admin and guest accounts, because roles (1,2,3) were unique. When third
account was created, it was impossible to create more of them.

Without the unique constraint in table Role, I got errors: "There is
no unique constraint matching given keys for referenced table account".

Now there is still the unique constraint in table Role, column role:
"unique=True", but in other tables which have the role column as foreign
key, the unique constraint is set to False, and it works.